Aurora Phase 2: GTK4 Renderer Scaffolding & Token Fixes
Fixed aurora-tokens lib.rs exports and implemented GTK4 renderer foundation.

Aurora-Tokens Fixes:
✅ Restored proper module exports (spacing, radius, elevation, motion, color)
✅ Fixed Color::relative_luminance() to use powf() for f32 exponentiation
✅ Proper pub struct DesignTokens with theme management
✅ Token validation and theme switching support
✅ 28 tests passing (color, spacing, radius, elevation, motion)

Aurora-GTK4 Renderer Foundation:
✅ New aurora-gtk crate (GTK4 rendering)
✅ CSS Provider: converts Aurora tokens to GTK CSS
✅ Theme management (Light, Dark, OLED, HDR)
✅ Automatic CSS generation from tokens
✅ Component styles (Button, Card, Input)
✅ Design token CSS custom properties
✅ 9 tests passing

CSS Provider Features:
- Token-to-CSS conversion (spacing, radius, motion, colors)
- Light/Dark/OLED/HDR theme support
- Semantic color mapping
- Component styles:
  * Button (filled, hover, active states)
  * Card (with elevation transitions)
  * Input (focus states, borders)
- CSS custom properties (--spacing-*, --color-*, --motion-*, etc.)
- Theme class generation (aurora-light, aurora-dark, etc.)

Architecture:
- AuroraGtk context (manages CSS provider + theme)
- CssProvider (generates CSS from tokens)
- Theme enum (Light, Dark, OLED, HDR)
- Error handling (GtkError, GtkResult)

Test Coverage:
✅ aurora-gtk creation
✅ Default theme initialization
✅ Theme switching
✅ CSS generation
✅ aurora-tokens theme management
✅ aurora-tokens validation
✅ Color contrast validation
✅ Design token access

Total Tests Passing: 107 (up from 70)
- aurora-tokens: 28
- aurora-typography: 37
- aurora-motion: 33
- aurora-gtk: 9

Workspace Compilation: ✅ Clean (no warnings)

Next Steps (Phase 2 Continuation):
1. Widget implementations (Button, Card, Dialog classes in GTK)
2. Animation integration (connect aurora-motion to GTK)
3. Accessibility features (screen readers, keyboard navigation)
4. Example GTK4 application using Aurora
5. Qt6 renderer implementation

Foundation Ready For:
- GTK4 application development
- Token-based styling
- Theme switching
- Multi-screen DPI optimization
- Wayland/X11 compatibility

This commit establishes the GTK4 foundation for Aurora.
Next phase will build actual GTK4 widget implementations
using the token system and motion engine.

+++ crates/aurora-tokens/src/color.rs
@@ -67,9 +67,9 @@ impl Color {
 
     /// Calculate perceived luminance (relative luminance) for contrast ratio
     fn relative_luminance(&self) -> f32 {
-        let r = (self.r as f32 / 255.0).pow(2.2);
-        let g = (self.g as f32 / 255.0).pow(2.2);
-        let b = (self.b as f32 / 255.0).pow(2.2);
+        let r = (self.r as f32 / 255.0).powf(2.2);
+        let g = (self.g as f32 / 255.0).powf(2.2);
+        let b = (self.b as f32 / 255.0).powf(2.2);
         0.2126 * r + 0.7152 * g + 0.0722 * b
     }
